service get / delete / add

+++ app/src/main/java/com/example/ibrasaloonapp/domain/model/Service.kt
@@ -1,9 +1,8 @@
 package com.example.ibrasaloonapp.domain.model
 
-import com.google.gson.annotations.SerializedName
 
 data class Service(
     val id: String? = null,
     val title: String,
-    val price: Int,
+    val price: Double,
 )
